### 2021-11-24 #48 @elboyran
eXplaianble tool for scientists - Deep Insight And Neural Networks Analysis (DIANNA)
Submitting Author: Elena Ranguelova (@elboyran)
Package Name: dianna
One-Line Description of Package: an eXplainable AI (XAI) python library targeted at scientists
Repository Link (if existing):   https://github.com/dianna-ai/dianna

---

Modern scientific challenges are often tackled with (Deep) Neural Networks (DNN). Despite their high predictive accuracy, DNNs lack inherent explainability. Many DNN users, especially scientists, do not harvest DNNs power because of lack of trust and understanding of their working.

Meanwhile, the eXplainable AI (XAI) methods offer some post-hoc interpretability and insight into the DNN reasoning. This is done by quantifying the relevance of individual features (image pixels, words in text, etc.) with respect to the prediction. These "relevance heatmaps" indicate how the network has reached its decision directly in the input modality (images, text, speech etc.) of the data.

There are many Open Source Software (OSS) implementations of these methods, alas, supporting a single DNN format and the libraries are known mostly by the AI experts. The DIANNA library supports the best XAI methods in the context of scientific usage providing their OSS implementation based on the ONNX standard and demonstrations on benchmark datasets. Representing visually the captured knowledge by the AI system can become a source of (scientific) insights.

It is work in progress, for now DIANNA supports part of chosen by objective criteria AI explainability methods such as RISE, LIME
and DeepLIFT SHAP (under development) for ONNX models, which means it's AI framework agnostic.

A possible category could be Data visualization, but the actual one is missing from the list. It would fall under the broad one of Data Analytics and more precisely eXplainable AI.

Scientists in any domain (especially not (X)AI experts), but also any other AI users who want to open the AI "black boxes". Very much so also for the XAI developers who want to study and compare against the proposed benchmarks we offer, the properties of their methods to the state-of-the-art ones. The scientific application potential is enormous and not limited to any science domain. Examples are given , e,g, in these publications:
[Explainable Machine Learning for Scientific Insights and Discoveries](https://arxiv.org/pdf/1905.08883.pdf)
[Explaining Deep Neural Networks and Beyond: A Review of Methods and Applications](https://ieeexplore.ieee.org/stamp/stamp.jsp?tp=&arnumber=9369420), nice summary illustration in Fig. 20 on page 268

As mentioned in the Description, there are many packages (shap, lime, etc.) and even libraries (e.g. Captum, iNNvestigate, etc.) implementing eith a single XAI method or a group of methods (without a clearly motivated choice, than own research) for a single DNN format (e.g. only pytorch or keras (tensorflow)) . With the careful objective selection of  XAI methods and the choice of ONNX standard supported, it makes it the only library applicable to any trained AI model, independent of the framework, hence espacialaly useful for any domain scientists. We aim to support also many data domains- images, text, and in the future time-series, tabular and graph data.

It is work in progress, but we are already thinking of how to disseminate it. We want to reach and be useful to as many domains scientists  as possible. That is the mission of our organization- the [Netherlands eScience center](nlesc.nl).

**The question is does it fit in the scope of pyOpenSci?**
