### 2021-11-30 e795ccf
shuffle standup name order
