+++ meeting_templates/standup-checklist.md
@@ -22,7 +22,7 @@ We aim at finishing the stand up in 25 mins. Longer discussions should be held i
 
 1. Discuss:
    1. [Issues with "standup" labeled](https://github.com/dianna-ai/dianna/issues?q=is%3Aissue+is%3Aopen+label%3Astandup)
-   1. [pull requests](https://github.com/dianna-ai/dianna/pulls)
+   1. [pull requests dianna](https://github.com/dianna-ai/dianna/pulls); [pull requests dianna-exploration](https://github.com/dianna-ai/dianna-exploration/pulls)
    1. Does anyone want to do pair programming? Make an appointment.
 1. Any other business, e.g.
    - Issues unrelated to the sprint
