### 2021-12-01 #48 @elboyran
Hello, I was wondering if this pre-submission form has been noticed?
For me, it is important to find out **if my software fits the scope of pyOpenSci**.
