### 2021-12-01 #48 @NickleDave
Hi @elboyran and welcome to the pyOpenSci community.
Thank you for providing a detailed pre-inquiry submission and stating a very clear question.

@lwasser and I have discussed whether the package is in scope.
(Sorry for not getting back to you sooner -- we just had a holiday here in the US. Which shows you that we need a better distribution of editors 🙂)

The short answer is: yes, we think DIANNA could be in scope, but we do have a couple of questions.

We are still learning about how we can best support the scientific community using Python. It's clear from your inquiry and others like it that there's a need for tooling that supports researchers using deep learning methods. However, these kind of tools are not yet as common in the [rOpenSci](https://ropensci.org/software-review/) space, and we have modeled ourselves on that community. So we are figuring this out as we go.

Here are the areas of scope where we think the goals of DIANNA fit into pyOpenSci:
- data visualization
- reproducibility
- data extraction

I'll explain each.

#### data visualization
This one you identified and is fairly self-evident, although reasonable people could disagree about whether XAI methods are really visualizing data per se. But no one can deny that scientists in particular rely on visualizations to validate and tune algorithms they apply to data. So I think I would call this one "clearly in scope".

#### reproducibility
As you state, one of your goals is to replicate existing methods, with the overall goal of assisting domain scientists. The additional goal of achieving this through the ONXX standard would also increase reproducibility / replicability.

There is precedent for this being in scope: we previously provided review for `pystiche` which similarly encapsulates deep learning-based style transfer methods with the goal of standardizing research in that area.

https://github.com/pyOpenSci/software-review/issues/25

#### data extraction
I think meeting the ONXX standard could fall under this heading; you would make it possible for researchers to use XAI methods in a framework-agnostic manner, if I am understanding correctly. So DIANNA would allow researchers to "extract data" (trained weights) from models.

So, with that said, here are our questions:
- Please say more about what specifically DIANNA will provide that other XAI libraries do not.
  How will DIANNA and the community around it give scientists additional guidance? E.g., if I am a researcher and I can train a CNN model, then I am probably competent enough to apply GRAD-CAM to a layer in my CNN. I don't need an extra library to do that. One way I could image DIANNA augmenting this would be for the docs to consist of vignettes, where each vignette describes a use case for a specific domain. But I do not see that in the library now (or any docs, just let me know if I am missing them.) Please if you can provide more details about this.

- How close would you say DIANNA is to being ready-for-use by researchers?
  * we ask that any tool/library we review be near "maturing" state [as defined here in ROpenSci](https://devguide.ropensci.org/authors-guide.html)
    + see specifically the language on ROpenSci guide:
       > For any submission or pre-submission inquiry the README of your package should provide enough information about your package (goals, usage, similar packages) for the editors to assess its scope without having to install the package.
       > At the submission stage, all major functions should be stable enough to be fully documented and tested; the README should make a strong case for the package (the editors will read it to e.g. evaluate this as in scope or not.).

     I can see you all are working hard on development, but I don't see anything in the README that tells me how or when to use the library, e.g. a minimal snippet that demonstrates usage.

  * the reason we have this requirement is because our completed review represents a sign of approval on the API, and we want to avoid giving this approval to a library that then drastically changes its API after the review process
  * WRT the goal of providing interoperability via ONXX, my understanding is that this is far from trivial, as suggested by discussion on this issue: https://github.com/dianna-ai/dianna/issues/39

- How comfortable would you be with review from developers of other XAI libraries?
  Those developers seem like the most likely candidates to provide informative reviews (along with researchers from specific domains you address). We would like to make sure you are ready to engage with that community of developers before we invite them to review. Alternatively, you may be operating in more of an "open stealth" mode. Please let us know.

So overall we do feel that DIANNA could be in scope and we would be very interested in providing review.
We also definitely want to connect with your organization since we clearly have aligned goals.

As of right now, though, I think we would need to see the requirement met that the package is near a "maturing" state,  before initiating a review. Unless @lwasser feels very strongly otherwise.

Maybe it would be best to contact us again when you feel you are nearing that state?
Happy to discuss further here though.
