### 2021-12-08 9317804
refactor: isolate reading of p_keep variable
