### 2021-12-08 #48 @elboyran
Dear David (@NickleDave) and @lwasser,

Thank you very much for your time and detailed feedback.

First the questions covered by short answers:

- Yes, DIANNA is not mature enough and we are very much aware of that. I  have approached you via a pre-submission enquiry as the suggested means to find out if the software will fit into the scope of pyOpenSci in advance.
- When the time comes for proper submission, we are fine for other XAI developers to review.

Data visualization and in a way reproducibility are matching categories. I am not sure if the Data extraction is a good category and maybe you would agree with me after my answers to your questions.

- **How DIANNA is different than others.** Our target users are indeed researchers, but not (necessarily) ones who can develop and even train DNNs themselves. In our center we work with lots of as we call them “domain scientists” being researchers in almost any scientific domain (except from  computer science, though we also work with them, but on different level). We work collaboratively in projects where they lead the domain science research and we offer them a lot of ICT expertise - from basic software engineering best practices to HPC, AI and data analytics, we make research software for people working in biology, chemistry, physics, astronomy, social sciences and humanities, archaeology etc. Therefore we assume no prior knowledge in (X)AI (including specific AI frameworks, formats or XAI methods and libraries ), but want to address possible (and natural especially for scientists)  skepticism, towards using  “deep black boxes” and XAI methods on top, who want to explain the black boxes. An archeologist researching Via Appia in Rome, historian journalist, a social scientist studying slums, a biologist interested in the habitat of some bird species who found or were given (or created/trained themselves of course)  a fantastic trained ML/DL model is some format, etc. should trust and benefit (potentially even getting scientific insights) from our library for their research with none/limited X(AI) knowledge. In order to achieve that we aimed for 3 main developments which we did not find out in the current libraries:

1. Using single standardized ML model format- ONNX. All (including the big) libraries support mostly a single format- pytorch, TF etc. Our scientists should be free of the worry to be finding a library (not all libraries implement “all” XAI methods) supporting their format, let alone that they could possess many different format models for their research. In additon, ONNX has been created not only for interoperability in mind, but also performance. If people use ONNX models, they will benefit from that and will need an explainable library.
2. It seems a bit arbitrary which models are implemented/supported by each existing library. We base our choice on objective evaluation criteria. We also try to have “complementary” (e.g. with and without access to model architecture) XAI methods included.
3. DIANNA is not only software. It comes with benchmark datasets which we have designed in attempt to fight  the “human seems-to-make-sense” kind of bias. For me, conclusions in publications summarizing as “quantitatively our method works fine as it highlights the object’s edges” on big  image datasets is not convincing at all if my scientific data (e.g.. satellite image of an Indian city possibly containing a slum area I want to find out) looks nothing like ImageNet. That’s why we think experimenting with simple (in complexity terms) datasets can make scientist understand and trust the XAI on intuitive level in order to trust the AI. These benchmarks could also be useful for studying the properties of XAI methods by their creators.
Easy interface, plenty of documentation, tutorial etc. should make it easier for the (non-technical) users too.

Some of the work explanations above have happened in a [sibling repo](https://github.com/dianna-ai/dianna-exploration) (we are considering in merging them), but all motivation and plan is described in my initial funding proposal, which could send if interested.

I am not sure I understand “So DIANNA would allow researchers to "extract data" (trained weights) from models.” Doesn’t any (X)AI library allow that? Also, imho, this is no the main function of a XAI library.

Will be glad to hear your opinion on the scope given these extra explanations. Is it possible to expand your list (as your repo documentation suggests) and include e.g. data analytics?

Also, please, send me your thoughts on how can we connect on organizational level.
