### 2021-12-14 a52792a
allow users to pass all possible kwargs to LIME
