### 2021-12-14 9abc040
WIP add failing p_keep autotuning test (refs #24)
