+++ dianna/methods/rise.py
@@ -88,7 +88,10 @@ class RISE:
         # data shape without batch axis and (optional) channel axis
         img_shape = input_data.shape[1:3]
         p_keep = self._determine_p_keep_for_images(input_data, batch_size, runner)
-        self.masks = self.generate_masks_for_images(img_shape, p_keep)  # Expose masks for to make user inspection possible
+
+        # Expose masks for to make user inspection possible
+        self.masks = self.generate_masks_for_images(img_shape, p_keep, self.n_masks)
+
         masked = input_data * self.masks
 
         predictions = []
@@ -100,37 +103,38 @@ class RISE:
         return normalize(saliency, self.n_masks, p_keep)
 
     def _determine_p_keep_for_images(self, input_data, batch_size, runner):
-        runner = get_function(runner, preprocess_function=self.preprocess_function)
-
-        # data shape without batch axis and (optional) channel axis
+        p_keep = 0.5
         img_shape = input_data.shape[1:3]
-        p_keep = 0.5  # self._determine_p_keep_for_images(input_data, batch_size, runner)
-        self.masks = self.generate_masks_for_images(img_shape, p_keep)  # Expose masks for to make user inspection possible
-        masked = input_data * self.masks
+        masks = self.generate_masks_for_images(img_shape, p_keep, self.n_masks)
+        masked = input_data * masks
 
         predictions = []
         for i in tqdm(range(0, self.n_masks, batch_size), desc='Explaining'):
             predictions.append(runner(masked[i:i + batch_size]))
         predictions = np.concatenate(predictions)
-        self.predictions = predictions
-        saliency = predictions.T.dot(self.masks.reshape(self.n_masks, -1)).reshape(-1, *img_shape)
-        return normalize(saliency, self.n_masks, p_keep)
-        # p_keep = 0.5
+        print('predictions.shape', predictions.shape, 'std', predictions.std(axis=0))
+        return 0.5
+
+        # p_keeps = [0.5]
+        # n_masks = 50
         # img_shape = input_data.shape[1:3]
-        # masks = self.generate_masks_for_images(img_shape, p_keep)
-        # masked = input_data * masks
         #
-        # predictions = []
-        # for i in tqdm(range(0, self.n_masks, batch_size), desc='Explaining'):
-        #     predictions.append(runner(masked[i:i + batch_size]))
-        # predictions = np.concatenate(predictions)
+        # for p_keep in p_keeps:
+        #     masks = self.generate_masks_for_images(img_shape, p_keep, n_masks)
+        #     masked = input_data * masks
+        #
+        #     predictions = []
+        #     for i in tqdm(range(0, self.n_masks, batch_size), desc='Explaining'):
+        #         predictions.append(runner(masked[i:i + batch_size]))
+        #     predictions = np.concatenate(predictions)
+        #     print('predictions.shape', predictions.shape, 'std', predictions.std(axis=0))
         # return 0.5
 
 
     def _determine_p_keep(self):
         return self.p_keep if not self.p_keep is None else 0.5
 
-    def generate_masks_for_images(self, input_size, p_keep):
+    def generate_masks_for_images(self, input_size, p_keep, n_masks):
         """Generate a set of random masks to mask the input data
 
         Args:
@@ -145,9 +149,9 @@ class RISE:
                                 p=(p_keep, 1 - p_keep))
         grid = grid.astype('float32')
 
-        masks = np.empty((self.n_masks, *input_size))
+        masks = np.empty((n_masks, *input_size))
 
-        for i in tqdm(range(self.n_masks), desc='Generating masks'):
+        for i in tqdm(range(n_masks), desc='Generating masks'):
             y = np.random.randint(0, cell_size[0])
             x = np.random.randint(0, cell_size[1])
             # Linear upsampling and cropping
