### 2021-12-14 3534a8c
Move testing and coverage to pyproject.toml + add tox config
By making testing and coverage separate parts of the configuration, VSCode no longer
chokes on discovering tests. Hopefully, this will also make test
discovery easier in IDEs like PyCharm. I think the main problem was that
coverage was being treated as default part of testing by always passing
coverage flags to pytest. Moving this to pyproject.toml may have been
unnecessary in the end, but the resources I found online suggested this
was the way to go.
