+++ setup.cfg
@@ -60,16 +60,13 @@ dev =
     sphinx_rtd_theme
     sphinx-autoapi
     torchtext
+    coverage [toml]
 publishing =
     twine
     wheel
 
 [options.packages.find]
 include = dianna, dianna.*
-
-[coverage:run]
-branch = True
-source = dianna
 
 [isort]
 lines_after_imports = 2
@@ -79,8 +76,3 @@ known_first_party = dianna
 src_paths = dianna,tests
 line_length = 120
 
-[tool:pytest]
-testpaths = tests
-# Note that visual debugger in some editors like pycharm gets confused by coverage calculation.
-# As a workaround, configure the test configuration in pycharm et al with a --no-cov argument
-addopts = --cov --cov-report xml --cov-report term --cov-report html
