### 2021-12-15 9d86601
add docstring linting + remove deprecated isort flag
