### 2021-12-15 5c97429
add --diff to isort (suggests fixes instead of just complaining)
