### 2021-12-15 1fc9973
remove python 3.6 from setup.cfg and build matrix
