### 2021-12-16 b2f6fb5
Fix shape of output in rise test; closes #62
