### 2021-12-16 7fb9c57
No longer convert data to float32 in SimpleModelRunner; this is up to the user now that we keep track of dtypes internally
