+++ dianna/utils/onnx_runner.py
@@ -1,4 +1,3 @@
-import numpy as np
 import onnxruntime as ort
 
 
@@ -29,6 +28,6 @@ class SimpleModelRunner:
         if self.preprocess_function is not None:
             input_data = self.preprocess_function(input_data)
 
-        onnx_input = {input_name: input_data.astype(np.float32)}
+        onnx_input = {input_name: input_data}
         pred_onnx = sess.run([output_name], onnx_input)[0]
         return pred_onnx
