### 2021-12-16 c3fd6d7
Add utility to move an xarray axis and use it in RISE
