+++ dianna/utils/__init__.py
@@ -1,2 +1,2 @@
 # flake8: noqa: F401
-from .misc import get_function, to_xarray
+from .misc import get_function, to_xarray, move_axis
