### 2021-12-16 09bd787
Add tests for xarray manipulation
