+++ dianna/utils/__init__.py
@@ -1,2 +1,4 @@
 # flake8: noqa: F401
-from .misc import get_function, to_xarray, move_axis
+from .misc import get_function
+from .misc import move_axis
+from .misc import to_xarray
