+++ setup.cfg
@@ -41,6 +41,7 @@ install_requires =
     onnxruntime
     Ipython
     matplotlib
+    xarray
 
 [options.data_files]
 # This section requires setuptools>=40.6.0
