### 2021-12-16 #48 @NickleDave
Hi @elboyran and thank you for your clear point-by-point response.

Everything you are saying makes total sense to me.
I think we agree that potentially some aspects of DIANNA are in scope for pyOpenSci, under the categories of visualization and reproducibility.

However, another factor we would need to consider is that XAI is an active research area. I'm sure you are aware of this, as your statement about "skeptical scientists" makes clear.

We would really need to make sure we have the right reviewers, that could address questions about XAI methods, especially with respect to fairness, bias, interpretability and usability.

I do not think we can provide you with this kind of review at this time. As we state in our guide:
> Note that we cannot currently accept statistics and/or machine learning packages. We don’t feel that we can give these the deep, thorough review that they require.

You are also right that this would be venturing into the territory of "data analytics". I did discuss this with @lwasser who also felt that adding an analytics category would go beyond our intended scope. We are following the ROpenSci model where our central focus is on ["tooling for the data life cycle"](https://devguide.ropensci.org/policies.html?q=scope#aims-and-scope). That community has only very recently added statistics packages. As a relatively new organization, we are just not able to extend our scope the same way.

You may be better served by going through review at a traditional journal.

Please let me know how that sounds. I am happy to discuss with you further.
