### 2021-12-20 c52578d
Fix linting issue and add test for moving non-existent axis
