+++ dianna/utils/__init__.py
@@ -1,5 +1,5 @@
 # flake8: noqa: F401
 from .misc import get_function
+from .misc import get_kwargs_applicable_to_function
 from .misc import move_axis
 from .misc import to_xarray
-from .misc import get_kwargs_applicable_to_function
