### 2021-12-21 38fb757
Implement use of xarray for LIME on images Referene test output is updated as original test contained unneeded axis swap in the input data
