+++ tests/test_lime.py
@@ -11,12 +11,11 @@ class LimeOnImages(TestCase):
 
     def test_lime_function(self):
         np.random.seed(42)
-        # shape is batch, y, x, channel
         input_data = np.random.random((1, 224, 224, 3))
+        labels = ('batch', 'y', 'x', 'channels')
 
-        explainer = LIME(random_state=42)
+        explainer = LIME(random_state=42, axes_labels=labels)
         heatmap = explainer.explain_image(run_model, input_data, num_samples=100)
-
         heatmap_expected = np.load('tests/test_data/heatmap_lime_function.npy')
         assert heatmap.shape == input_data[0].shape[:2]
         assert np.allclose(heatmap, heatmap_expected, atol=.01)
@@ -24,17 +23,23 @@ class LimeOnImages(TestCase):
     def test_lime_filename(self):
         np.random.seed(42)
         model_filename = 'tests/test_data/mnist_model.onnx'
-        black_and_white = generate_data(batch_size=1).transpose((0, 3, 2, 1))
-        input_data = np.zeros(list(black_and_white.shape[:-1]) + [3]) + black_and_white
+
+        # batch channel y x
+        black_and_white = generate_data(batch_size=1)
+        # Make data 3-channel instead of 1-channel
+        input_data = np.zeros([1, 3] + list(black_and_white.shape[2:])) + black_and_white
+        input_data = input_data.astype(np.float32)
+        labels = ('batch', 'channels', 'y', 'x')
 
         def preprocess(data):
-            data = data[..., 0][..., None]
-            return np.moveaxis(data, -1, 1).astype(np.float32)
+            # select single channel out of 3, but keep the channel axis
+            return data[:, [0], ...]
 
-        heatmap = dianna.explain_image(model_filename, input_data, method="LIME", preprocess_function=preprocess, random_state=42)
+        heatmap = dianna.explain_image(model_filename, input_data, method="LIME", preprocess_function=preprocess, random_state=42,
+                                       axes_labels=labels)
 
         heatmap_expected = np.load('tests/test_data/heatmap_lime_filename.npy')
-        assert heatmap.shape == input_data[0].shape[:2]
+        assert heatmap.shape == input_data[0, 0].shape
         assert np.allclose(heatmap, heatmap_expected, atol=.01)
 
 
