+++ dianna/methods/lime.py
@@ -1,6 +1,7 @@
 from lime.lime_image import LimeImageExplainer
 from lime.lime_text import LimeTextExplainer
 from dianna import utils
+import numpy as np
 
 
 class LIME:
@@ -19,6 +20,7 @@ class LIME:
                  mask_string=None,
                  random_state=None,
                  char_level=False,
+                 axes_labels=None,
                  preprocess_function=None,
                  ):  # pylint: disable=too-many-arguments
         """LIME initializer.
@@ -34,6 +36,9 @@ class LIME:
             mask_string (str, optional):
             random_state (int or np.RandomState, optional):
             char_level (bool, optional):
+            axes_labels (dict/list, optional): If a dict, key,value pairs of axis index, name.
+                                               If a list, the name of each axis where the index
+                                               in the list is the axis index
             preprocess_function (callable, optional): Function to preprocess input data with
         """
         self.text_explainer = LimeTextExplainer(kernel_width,
@@ -56,6 +61,8 @@ class LIME:
                                                   )
 
         self.preprocess_function = preprocess_function
+        self.axes_labels = axes_labels if axes_labels is not None else []
+        self.required_labels = ('batch', 'channels')
 
     def explain_text(self,
                      model_or_function,
@@ -111,7 +118,7 @@ class LIME:
                       positive_only=False,
                       hide_rest=True,
                       **kwargs,
-                      ):  # pylint: disable=too-many-arguments
+                      ):  # pylint: disable=too-many-arguments,too-many-locals
         """
         Run the LIME explainer.
 
@@ -129,12 +136,10 @@ class LIME:
         Returns:
             list of (word, index of word in raw text, importance for target class) tuples
         """
-        runner = utils.get_function(model_or_function, preprocess_function=self.preprocess_function)
-        # remove batch axis from input data; this is only here for a consistent API
-        # but LIME wants data without batch axis
-        if not len(input_data) == 1:
-            raise ValueError("Length of batch axis must be one.")
-        input_data = input_data[0]
+        input_data, full_preprocess_function = self._prepare_image_data(input_data)
+        runner = utils.get_function(model_or_function, preprocess_function=full_preprocess_function)
+
+        # run the explanation.
         explain_instance_kwargs = utils.get_kwargs_applicable_to_function(self.image_explainer.explain_instance, kwargs)
         explanation = self.image_explainer.explain_instance(input_data,
                                                             runner,
@@ -149,3 +154,48 @@ class LIME:
         mask = explanation.get_image_and_mask(label, positive_only=positive_only, hide_rest=hide_rest,
                                               num_features=num_features, **get_image_and_mask_kwargs)[1]
         return mask
+
+    def _prepare_image_data(self, input_data):
+        """
+        Transforms the data to be of the shape and type LIME expects
+
+        Args:
+            input_data (np.array): Data to be explained
+
+        Returns:
+            transformed input data, preprocessing function to use with utils.get_function()
+        """
+        input_data = utils.to_xarray(input_data, self.axes_labels, self.required_labels)
+        # remove batch axis from input data; this is only here for a consistent API
+        # but LIME wants data without batch axis
+        if not len(input_data['batch']) == 1:
+            raise ValueError("Length of batch axis must be one.")
+        input_data = input_data.sel(batch=0)
+        # ensure channels axis is last and keep track of where it was so we can move it back
+        channels_axis_index = input_data.dims.index('channels')
+        input_data = utils.move_axis(input_data, 'channels', -1)
+        # create preprocessing function that puts model input generated by LIME into the right shape and dtype,
+        # followed by running the user's preprocessing function
[…]
