### 2021-12-21 51a6b78
add requirements.txt at the top level for all dependencies
