+++ requirements.txt
@@ -1,0 +1,18 @@
+captum==0.4.x
+dianna==0.1.0
+keras==2.4.0
+lime==0.2.0.0
+matplotlib==3.4.2
+numpy==1.19.2
+onnx==1.9.0
+onnxruntime==1.7.2
+pandas==1.3.4
+scipy==1.6.2
+seaborn==0.11.2
+shap==0.39.0
+sklearn==1.0.1
+tensorflow==2.4.1
+torch==1.8.1
+torchtext==0.9.1
+torchvision==0.9.1
+tqdm==4.62.3
