### 2021-12-22 cb8ec88
autotune p_keep for images (refs #24)
