### 2021-12-22 7e8b498
fix test: remove forgotten line in test code
