### 2021-12-22 5684ac0
Fix prospector E226 warnings (spaces around operators)
Weirdly, these only show up when I run prospector on my laptop, not with
prospector on CI... I use the same version of prospector too, so not
sure what's going on.
