+++ dianna/methods/lime.py
@@ -197,5 +197,5 @@ class LIME:
         # axis in the right place
         # one is added to the channels axis index because there is an extra first axis: the batch axis
         if self.preprocess_function is None:
-            return lambda data: np.moveaxis(data, -1, channel_axis_index+1).astype(dtype)
-        return lambda data: self.preprocess_function(np.moveaxis(data, -1, channel_axis_index+1).astype(dtype))
+            return lambda data: np.moveaxis(data, -1, channel_axis_index + 1).astype(dtype)
+        return lambda data: self.preprocess_function(np.moveaxis(data, -1, channel_axis_index + 1).astype(dtype))
