### 2021-12-22 a854749
Fix linter and isort warnings
