+++ tests/test_lime.py
@@ -4,7 +4,8 @@ import dianna
 import dianna.visualization
 from dianna.methods import LIME
 from tests.test_onnx_runner import generate_data
-from tests.utils import ModelRunner, run_model
+from tests.utils import ModelRunner
+from tests.utils import run_model
 
 
 class LimeOnImages(TestCase):
