### 2021-12-22 8c397ec
fix linting (spaces around operators)
