### 2021-12-22 430a022
add autotune p_keep for text (refs #24)
