### 2021-12-22 f276ae0
fix formatting and linter suggestions
