+++ dianna/utils/onnx_runner.py
@@ -1,5 +1,5 @@
-import onnxruntime as ort
 import numpy as np
+import onnxruntime as ort
 
 
 class SimpleModelRunner:
