### 2021-12-23 d9f331f
Fix rise data types and remove comments from test (moved to issue)
