### 2021-12-23 3d2518f
add n_mask default value explanation
