+++ dianna/__init__.py
@@ -7,7 +7,7 @@ logging.getLogger(__name__).addHandler(logging.NullHandler())
 
 __author__ = "DIANNA Team"
 __email__ = "dianna-ai@esciencecenter.nl"
-__version__ = "0.1.0"
+__version__ = "0.2.0"
 
 
 def explain_image(model_or_function, input_data, method, **kwargs):
