+++ setup.cfg
@@ -25,7 +25,7 @@ name = dianna
 project_urls =
     Bug Tracker = https://github.com/dianna-ai/dianna/issues
 url = https://github.com/dianna-ai/dianna
-version = 0.1.0
+version = 0.2.0
 
 [options]
 zip_safe = False
