### 2021-12-23 5101249
Use only best class for auto-tuning of p_keep
