+++ dianna/methods/rise.py
@@ -75,9 +75,9 @@ class RISE:
         for i in range(0, n_masks, batch_size):
             current_input = masked[i:i + batch_size]
             current_predictions = runner(current_input)
-            predictions.append(current_predictions)
+            predictions.append(current_predictions.max(axis=1))
         predictions = np.concatenate(predictions)
-        std_per_class = predictions.std(axis=0)
+        std_per_class = predictions.std()
         return np.mean(std_per_class)
 
     def _generate_masks_for_text(self, input_shape, p_keep, n_masks):
@@ -123,15 +123,12 @@ class RISE:
         Returns:
             Explanation heatmap for each class (np.ndarray).
         """
-
-        runner = utils.get_function(model_or_function, preprocess_function=self.preprocess_function)
         # convert data to xarray
         input_data = utils.to_xarray(input_data, self.axes_labels, RISE.required_labels)
         # batch axis should always be first
         input_data = utils.move_axis(input_data, 'batch', 0)
-        # ensure channels axis is last and keep track of where it was so we can move it back
-        channels_axis_index = input_data.dims.index('channels')
-        input_data = utils.move_axis(input_data, 'channels', -1)
+        input_data, full_preprocess_function = self._prepare_image_data(input_data)
+        runner = utils.get_function(model_or_function, preprocess_function=full_preprocess_function)
 
         p_keep = self._determine_p_keep_for_images(input_data, runner) if self.p_keep is None else self.p_keep
 
@@ -141,11 +138,7 @@ class RISE:
         self.masks = self.generate_masks_for_images(img_shape, p_keep, self.n_masks)
 
         # Make sure multiplication is being done for correct axes
-        masked = (input_data * self.masks)
-        # ensure channels axis is in original location again
-        masked = utils.move_axis(masked, 'channels', channels_axis_index)
-        # convert to numpy for onnx
-        masked = masked.values.astype(input_data.dtype)
+        masked = input_data * self.masks
 
         batch_predictions = []
         for i in tqdm(range(0, self.n_masks, batch_size), desc='Explaining'):
@@ -173,14 +166,14 @@ class RISE:
         batch_size = 50
         img_shape = input_data.shape[1:3]
         masks = self.generate_masks_for_images(img_shape, p_keep, n_masks)
-        masked = (input_data * masks).astype(input_data.dtype)
+        masked = input_data * masks
         predictions = []
         for i in range(0, n_masks, batch_size):
             current_input = masked[i:i + batch_size]
             current_predictions = runner(current_input)
-            predictions.append(current_predictions)
+            predictions.append(current_predictions.max(axis=1))
         predictions = np.concatenate(predictions)
-        std_per_class = predictions.std(axis=0)
+        std_per_class = predictions.std()
         return np.mean(std_per_class)
 
     def generate_masks_for_images(self, input_size, p_keep, n_masks):
@@ -198,7 +191,7 @@ class RISE:
                                 p=(p_keep, 1 - p_keep))
         grid = grid.astype('float32')
 
-        masks = np.empty((n_masks, *input_size))
+        masks = np.empty((n_masks, *input_size), dtype=np.float32)
 
         for i in range(n_masks):
             y = np.random.randint(0, cell_size[0])
@@ -207,3 +200,41 @@ class RISE:
             masks[i, :, :] = _upscale(grid[i], up_size)[y:y + input_size[0], x:x + input_size[1]]
         masks = masks.reshape(-1, *input_size, 1)
         return masks
+
+    def _prepare_image_data(self, input_data):
+        """
+        Transforms the data to be of the shape and type RISE expects
+
+        Args:
+            input_data (xarray): Data to be explained
+
+        Returns:
+            transformed input data, preprocessing function to use with utils.get_function()
+        """
+        # ensure channels axis is last and keep track of where it was so we can move it back
[…]
