### 2021-12-23 ba916a9
set prospector to strict linting
Line length 127 is the average of 3*42 and 128, which are both nice
numbers.
