### 2021-12-23 c539026
Add mypy to linting
Related to #85
