+++ .prospector.yml
@@ -28,3 +28,8 @@ pep257:
         D213,  # Multi-line docstring summary should start at the second line
         D404,  # First word of the docstring should not be This
     ]
+
+mypy:
+    run: true
+    options:
+        ignore-missing-imports: true
