+++ setup.cfg
@@ -51,6 +51,7 @@ install_requires =
 dev =
     bump2version
     prospector[with_pyroma]
+    mypy
     isort
     pytest
     pytest-cov
