### 2021-12-24 711ae9b
Add testing of tutorial notebooks using nbmake
