+++ requirements.txt
@@ -1,4 +1,4 @@
-captum==0.4.x
+captum==0.4.0
 dianna==0.1.0
 keras==2.4.0
 lime==0.2.0.0
