+++ requirements.txt
@@ -5,7 +5,7 @@ lime==0.2.0.0
 matplotlib==3.4.2
 numpy==1.19.2
 onnx==1.9.0
-onnxruntime==1.7.2
+onnxruntime==1.7.0
 pandas==1.3.4
 scipy==1.6.2
 seaborn==0.11.2
