+++ requirements.txt
@@ -10,7 +10,7 @@ pandas==1.3.4
 scipy==1.6.2
 seaborn==0.11.2
 shap==0.39.0
-sklearn==1.0.1
+scikit-learn==1.0.1
 tensorflow==2.4.1
 torch==1.8.1
 torchtext==0.9.1
