+++ requirements.txt
@@ -2,7 +2,6 @@ captum==0.4.0
 dianna==0.1.0
 keras==2.4.0
 lime==0.2.0.0
-matplotlib==3.4.2
 numpy==1.19.2
 onnx==1.9.0
 onnxruntime==1.7.0
