+++ requirements.txt
@@ -2,8 +2,8 @@ captum==0.4.0
 dianna==0.1.0
 keras==2.4.0
 lime==0.2.0.0
-numpy==1.19.2
 onnx==1.9.0
+onnx-tf==1.1.2
 onnxruntime==1.7.0
 pandas==1.3.4
 scipy==1.6.2
