### 2022-01-03 #48 @NickleDave
Closing this for now but again we are happy to discuss further
